message_processor/requirements.txt
redis==5.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncio==3.4.3 